requirements.txt
kiteconnect>=4.2.0
breeze-connect>=1.0.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pytz>=2024.1
ta>=0.11.0
